Move Notification into new file, and add convenience initializer
